server update
retour information sous format :
-18.8792/47.5079~30/35/20/15~site1?10/site2?37/site3?15/site4?5
Petit correction

+++ src/main/java/com/example/serveur/repository/StatusMessageRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StatusMessageRepository extends JpaRepository<StatusMessage, Long> {
+public interface StatusMessageRepository extends JpaRepository<StatusMessage, Integer> {
     @Query("SELECT s FROM StatusMessage s WHERE s.status = :status")
     StatusMessage findByStatus(@Param("status") String status);
 }
