+++ src/main/java/com/example/serveur/repository/TypeAlerteRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TypeAlerteRepository extends JpaRepository<TypeAlerte, Long> {
+public interface TypeAlerteRepository extends JpaRepository<TypeAlerte, Integer> {
     @Query("SELECT t FROM TypeAlerte t WHERE t.zone = :zone")
     TypeAlerte findByZone(@Param("zone") String zone);
 }
